move vectorDB, embed, model packages, and util.py to utils package.
