+++ KoPrivateGPT/utils/model/koalpaca.py
@@ -1,5 +1,5 @@
-from KoPrivateGPT.model.base_model_factory import BaseModelFactory
-from utils import StoppingCriteriaSub
+from KoPrivateGPT.utils.model.base_model_factory import BaseModelFactory
+from KoPrivateGPT.utils.util import StoppingCriteriaSub
 
 
 class KoAlpacaFactory(BaseModelFactory):
